Renames the base package
