+++ src/main/kotlin/nomic/game/manager/api/TestingSpring.kt
@@ -1,4 +1,4 @@
-package game.manager.nomic.api
+package nomic.game.manager.api
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
