+++ src/test/kotlin/nomic/game/manager/api/ApiApplicationTests.kt
@@ -1,4 +1,4 @@
-package game.manager.nomic.api
+package nomic.game.manager.api
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
